reboot improve

+++ ubuntu-preconfig.sh
@@ -235,8 +235,12 @@ Reboot(){
             [Yy]*) 
                 Show 4 "Preparing to reboot..."
                 # create a flag file to signal that we are resuming from reboot.
-                touch "$WORK_DIR/resume-after-reboot"
-                Check_Success "Flag file to resume after reboot"
+                if [ -f "$WORK_DIR"/resume-after-reboot ]; then
+                    touch "$WORK_DIR/resume-after-reboot"
+                    Check_Success "Flag file to resume after reboot"
+                else
+                    Show 0 "Flag file to resume after reboot success"
+                fi                
                 # add the link to bashrc to start the script on login
                 echo "curl -fsSL $SCRIPT_LINK | sudo bash" >> "$WORK_DIR"/.bashrc
                 Check_Success "Setting up run script on boot"
